Feature/sj

See merge request cloud_track/class_03/web_project2/team08/book-shopping-mall!66

+++ src/main/resources/static/admin-categories/admin-categories.js
@@ -8,6 +8,7 @@ document.addEventListener('DOMContentLoaded', () => {
     const categoryIdInput = document.getElementById('categoryId');
     const categoryNameInput = document.getElementById('categoryName');
     const deleteSelectedButton = document.getElementById('deleteSelected');
+    const errorMessage = document.getElementById('errorMessage');
 
     // Fetch all categories and display them
     function fetchCategories() {
@@ -45,18 +46,28 @@ document.addEventListener('DOMContentLoaded', () => {
         const method = categoryId ? 'PUT' : 'POST';
         const url = categoryId ? `/admin/category/${categoryId}` : '/admin/category';
 
+        errorMessage.innerText = '';
+
         fetch(url, {
             method: method,
             headers: {
                 'Content-Type': 'application/json'
             },
             body: JSON.stringify({ name: categoryName })
+        }).then(response => {
+            if (!response.ok) {
+                return response.json().then(error => {
+                    throw new Error(error.message || '동일한 카테고리명입니다. 다시 입력해주세요.');
+                });
+            }
+            return response.json();
         }).then(() => {
             categoryForm.reset();
             categoryIdInput.value = ''; // Reset the hidden input value
             fetchCategories(); // 변경된 데이터를 반영하기 위해 fetchCategories 호출
         }).catch(error => {
-            console.error('Error saving category:', error);
+            const errorMessage = document.getElementById('errorMessage');
+            errorMessage.innerText = error.message || '카테고리를 저장하는 동안 오류가 발생했습니다.';
         });
     });
 
@@ -68,28 +79,36 @@ document.addEventListener('DOMContentLoaded', () => {
 
     // Delete category
     window.deleteCategory = id => {
-        fetch(`/admin/category/${id}`, { method: 'DELETE' })
-            .then(() => fetchCategories()) // 변경된 데이터를 반영하기 위해 fetchCategories 호출
-            .catch(error => {
-                console.error('Error deleting category:', error);
-            });
+        if(confirm('정말로 이 카테고리를 삭제하시겠습니까?')) {
+            fetch(`/admin/category/${id}`, { method: 'DELETE' })
+                .then(() => fetchCategories()) // 변경된 데이터를 반영하기 위해 fetchCategories 호출
+                .catch(error => {
+                    console.error('Error deleting category:', error);
+                });
+        }
     };
 
     // Delete selected categories
     deleteSelectedButton.addEventListener('click', () => {
         const selectedIds = Array.from(document.querySelectorAll('.categoryCheckbox:checked'))
             .map(checkbox => checkbox.getAttribute('data-id'));
-        fetch('/admin/category', {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(selectedIds)
-        }).then(() => fetchCategories()) // 변경된 데이터를 반영하기 위해 fetchCategories 호출
-            .catch(error => {
-                console.error('Error deleting selected categories:', error);
-            });
+        if(confirm('정말로 선택한 카테고리들을 삭제하시겠습니까?')) {
+            fetch('/admin/category', {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(selectedIds)
+            }).then(() => fetchCategories()) // 변경된 데이터를 반영하기 위해 fetchCategories 호출
+                .catch(error => {
+                    console.error('Error deleting selected categories:', error);
+                });
+        }
     });
+
+    categoryNameInput.addEventListener('input', () => {
+        errorMessage.innerText = '';
+    })
 
     // Initial fetch of categories
     fetchCategories();
